Revert "Create custom hook for animation (#17)"
This reverts commit 5547432b965193347fb34ea4c362fe14ce84bcc5.
